+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,10 @@
         android:theme="@style/Theme.Bolt"
         android:networkSecurityConfig="@xml/network_security_config">
         <activity
-            android:name=".ProfileMenuLoggedOutActivity"
+            android:name=".AccountActivities.ProfileMenuLoggedOutActivity"
             android:exported="true" />
         <activity
-            android:name=".ProfileMenuLoggedInActivity"
+            android:name=".AccountActivities.ProfileMenuLoggedInActivity"
             android:exported="true" />
         <!--
              The API key for Google Maps-based APIs is defined as a string resource.
@@ -56,7 +56,7 @@
         </activity>
 
         <activity
-            android:name=".SampleActivity"
+            android:name=".payment.KlarnaActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW"/>
@@ -67,7 +67,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".OrderCompletedActivity"/>
+        <activity android:name=".payment.KlarnaOrderCompletedActivity"/>
     </application>
 
 </manifest>
