+++ app/src/main/java/com/flexicharge/bolt/payment/api/BasicAuthInterceptor.kt
@@ -1,4 +1,4 @@
-package com.flexicharge.bolt.api
+package com.flexicharge.bolt.payment.api
 
 import okhttp3.Credentials
 import okhttp3.Interceptor
