+++ app/src/main/res/layout/activity_profile_menu_logged_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileMenuLoggedInActivity">
+    tools:context=".AccountActivities.ProfileMenuLoggedInActivity">
 
     <ImageView
         android:id="@+id/imageView"
