+++ app/src/main/res/layout/activity_profile_menu_logged_out.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileMenuLoggedOutActivity">
+    tools:context=".AccountActivities.ProfileMenuLoggedOutActivity">
 
     <ImageView
         android:id="@+id/imageView"
